third_party/WebKit/LayoutTests/http/tests/xmlhttprequest/origin-whitelisting-https-expected.txt

CONSOLE WARNING: line 18: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
CONSOLE ERROR: line 20: Failed to load http://localhost:8000/xmlhttprequest/resources/get.txt?sync: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Failed to load http://localhost:8000/xmlhttprequest/resources/get.txt?async: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
Tests that origin whitelisting for https does not match http URLs.

Testing: http://localhost:8000/xmlhttprequest/resources/get.txt (sync)
PASS: NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://localhost:8000/xmlhttprequest/resources/get.txt?sync'.
Testing: http://localhost:8000/xmlhttprequest/resources/get.txt (async)
PASS: 0
